Removes sensor_constants
