+++ src/main.cpp
@@ -10,7 +10,6 @@
 
 #include <TeensyDebug.h>
 #include "sensors/LEDBoard.hpp"
-#include "sensor_constants.hpp"
 #include "SensorManager.hpp"
 
 #include "utils/watchdog.hpp"
