+++ src/sensors/Sensor.hpp
@@ -3,7 +3,17 @@
 #ifndef SENSOR_HPP
 #define SENSOR_HPP
 
-#include "sensor_constants.hpp"
+#include <stdint.h>
+
+enum SensorType : uint8_t {
+    BUFFENC = 0,
+    ICM = 1,
+    REVENC = 2,
+    TOF = 3,
+    LIDAR = 4,
+
+    // Add other sensor types here
+};
 
 /// @brief Base class for all sensors.
 class Sensor {
